Added delete content tests

+++ src/Inshapardaz.Domain/Ports/Handlers/Library/DeleteChapterContentRequest.cs
@@ -27,25 +27,31 @@ namespace Inshapardaz.Domain.Ports.Library
     public class DeleteChapterContentRequestHandler : RequestHandlerAsync<DeleteChapterContentRequest>
     {
         private readonly IChapterRepository _chapterRepository;
+        private readonly IFileRepository _fileRepository;
         private readonly IFileStorage _fileStorage;
 
-        public DeleteChapterContentRequestHandler(IChapterRepository chapterRepository, IFileStorage fileStorage)
+        public DeleteChapterContentRequestHandler(IChapterRepository chapterRepository, IFileRepository fileRepository, IFileStorage fileStorage)
         {
             _chapterRepository = chapterRepository;
+            _fileRepository = fileRepository;
             _fileStorage = fileStorage;
         }
 
         [Authorise(step: 1, HandlerTiming.Before)]
         public override async Task<DeleteChapterContentRequest> HandleAsync(DeleteChapterContentRequest command, CancellationToken cancellationToken = new CancellationToken())
         {
-            var contentUrl = await _chapterRepository.GetChapterContentUrl(command.LibraryId, command.BookId, command.ChapterId, command.Language, command.MimeType, cancellationToken);
+            var content = await _chapterRepository.GetChapterContent(command.LibraryId, command.BookId, command.ChapterId, command.Language, command.MimeType, cancellationToken);
 
-            if (!string.IsNullOrWhiteSpace(contentUrl))
+            if (content != null)
             {
-                await _fileStorage.TryDeleteFile(contentUrl, cancellationToken);
-            }
+                if (!string.IsNullOrWhiteSpace(content.ContentUrl))
+                {
+                    await _fileStorage.TryDeleteFile(content.ContentUrl, cancellationToken);
+                }
 
-            await _chapterRepository.DeleteChapterContentById(command.LibraryId, command.BookId, command.ChapterId, command.Language, command.MimeType, cancellationToken);
+                await _fileRepository.DeleteFile(content.FileId, cancellationToken);
+                await _chapterRepository.DeleteChapterContentById(command.LibraryId, command.BookId, command.ChapterId, command.Language, command.MimeType, cancellationToken);
+            }
 
             return await base.HandleAsync(command, cancellationToken);
         }
